+++ intTests/test_mir_verify_basic/test.rs
@@ -1,0 +1,51 @@
+pub fn id_array(x: [u32; 5]) -> [u32; 5] {
+    x
+}
+
+pub fn id_i8(x: i8) -> i8 {
+    x
+}
+
+pub fn id_i16(x: i16) -> i16 {
+    x
+}
+
+pub fn id_i32(x: i32) -> i32 {
+    x
+}
+
+pub fn id_i64(x: i64) -> i64 {
+    x
+}
+
+pub fn id_isize(x: isize) -> isize {
+    x
+}
+
+pub fn id_tuple(x: (u32, u32)) -> (u32, u32) {
+    x
+}
+
+pub fn id_u8(x: u8) -> u8 {
+    x
+}
+
+pub fn id_u16(x: u16) -> u16 {
+    x
+}
+
+pub fn id_u32(x: u32) -> u32 {
+    x
+}
+
+pub fn id_u64(x: u64) -> u64 {
+    x
+}
+
+pub fn id_usize(x: usize) -> usize {
+    x
+}
+
+pub fn id_unit(x: ()) -> () {
+    x
+}