+++ saw-remote-api/python/tests/saw/test-files/basic_mir.rs
@@ -1,0 +1,3 @@
+pub fn basic(x: u32) -> u32 {
+    x
+}