+++ src/SAWScript/Crucible/Common/Override.hs
@@ -20,7 +20,7 @@ Stability   : provisional
 
 module SAWScript.Crucible.Common.Override
   ( Pointer
-  , Pointer'
+  , MS.Pointer'
   , OverrideState
   , OverrideState'(..)
   , osAsserts
@@ -51,6 +51,7 @@ module SAWScript.Crucible.Common.Override
   , writeGlobal
   , failure
   , getSymInterface
+  , enforceCompleteSubstitution
   --
   , assignmentToList
   , MetadataMap
@@ -58,6 +59,7 @@ module SAWScript.Crucible.Common.Override
 
 import qualified Control.Exception as X
 import           Control.Lens
+import           Control.Monad (unless)
 import           Control.Monad.Trans.State hiding (get, put)
 import           Control.Monad.State.Class (MonadState(..))
 import           Control.Monad.Error.Class (MonadError)
@@ -66,7 +68,7 @@ import qualified Control.Monad.Fail as Fail
 import           Control.Monad.Trans.Except
 import           Control.Monad.Trans.Class
 import           Control.Monad.IO.Class
-import           Data.Kind (Type)
+import qualified Data.Map as Map
 import           Data.Map (Map)
 import           Data.Set (Set)
 import           Data.Typeable (Typeable)
@@ -93,6 +95,7 @@ import qualified What4.ProgramLoc as W4
 import           SAWScript.Exceptions
 import           SAWScript.Crucible.Common (Sym)
 import           SAWScript.Crucible.Common.MethodSpec as MS
+import           SAWScript.Crucible.Common.Setup.Value as MS
 
 -- TODO, not sure this is the best place for this definition
 type MetadataMap =
@@ -103,13 +106,11 @@ type MetadataMap =
 
 type LabeledPred sym = W4.LabeledPred (W4.Pred sym) Crucible.SimError
 
-type family Pointer' ext sym :: Type
-
-type Pointer ext = Pointer' ext Sym
+type Pointer ext = MS.Pointer' ext Sym
 
 data OverrideState' sym ext = OverrideState
   { -- | Substitution for memory allocations
-    _setupValueSub :: Map AllocIndex (Pointer' ext sym)
+    _setupValueSub :: Map AllocIndex (MS.Pointer' ext sym)
 
     -- | Substitution for SAW Core external constants
   , _termSub :: Map VarIndex Term
@@ -273,7 +274,7 @@ instance ( PP.Pretty (ExtType ext)
 
 instance ( PP.Pretty (ExtType ext)
          , PP.Pretty (MS.PointsTo ext)
-         , Typeable ext 
+         , Typeable ext
          ) => X.Exception (OverrideFailure ext)
 
 --------------------------------------------------------------------------------
@@ -375,6 +376,26 @@ failure loc e = OM (lift (throwE (OF loc e)))
 
 getSymInterface :: Monad m => OverrideMatcher' sym ext md m sym
 getSymInterface = OM (use syminterface)
+
+-- | Verify that all of the fresh variables for the given
+-- state spec have been "learned". If not, throws
+-- 'AmbiguousVars' exception.
+enforceCompleteSubstitution ::
+  W4.ProgramLoc ->
+  MS.StateSpec ext ->
+  OverrideMatcher ext w ()
+enforceCompleteSubstitution loc ss =
+
+  do sub <- OM (use termSub)
+
+     let -- predicate matches terms that are not covered by the computed
+         -- term substitution
+         isMissing tt = ecVarIndex (tecExt tt) `Map.notMember` sub
+
+         -- list of all terms not covered by substitution
+         missing = filter isMissing (view MS.csFreshVars ss)
+
+     unless (null missing) (failure loc (AmbiguousVars missing))
 
 ------------------------------------------------------------------------