+++ src/SAWScript/Crucible/JVM/Override.hs
@@ -109,6 +109,7 @@ import qualified SAWScript.Crucible.Common.Override as Ov (getSymInterface)
 import qualified SAWScript.Crucible.Common.MethodSpec as MS
 import           SAWScript.Crucible.JVM.MethodSpecIR
 import           SAWScript.Crucible.JVM.ResolveSetupValue
+import           SAWScript.Crucible.JVM.Setup.Value ()
 import           SAWScript.Options
 import           SAWScript.Panic
 import           SAWScript.Utils (handleException)
@@ -121,7 +122,6 @@ type SetupValue = MS.SetupValue CJ.JVM
 type CrucibleMethodSpecIR = MS.CrucibleMethodSpecIR CJ.JVM
 type StateSpec = MS.StateSpec CJ.JVM
 type SetupCondition = MS.SetupCondition CJ.JVM
-type instance Pointer' CJ.JVM Sym = JVMRefVal
 
 -- TODO: Improve?
 ppJVMVal :: JVMVal -> PP.Doc ann
@@ -354,24 +354,6 @@ learnCond opts sc cc cs prepost ss =
      enforceCompleteSubstitution loc ss
 
 
--- | Verify that all of the fresh variables for the given
--- state spec have been "learned". If not, throws
--- 'AmbiguousVars' exception.
-enforceCompleteSubstitution :: W4.ProgramLoc -> StateSpec -> OverrideMatcher CJ.JVM w ()
-enforceCompleteSubstitution loc ss =
-
-  do sub <- OM (use termSub)
-
-     let -- predicate matches terms that are not covered by the computed
-         -- term substitution
-         isMissing tt = ecVarIndex (tecExt tt) `Map.notMember` sub
-
-         -- list of all terms not covered by substitution
-         missing = filter isMissing (view MS.csFreshVars ss)
-
-     unless (null missing) (failure loc (AmbiguousVars missing))
-
-
 -- execute a pre/post condition
 executeCond ::
   Options ->
@@ -590,7 +572,7 @@ matchArg opts sc cc cs prepost md actual@(RVal ref) expectedTy setupval =
     MS.SetupNull () ->
       do sym <- Ov.getSymInterface
          p   <- liftIO (CJ.refIsNull sym ref)
-         addAssert p md (Crucible.SimError (cs ^. MS.csLoc) (Crucible.AssertFailureSimError ("null-equality " ++ stateCond prepost) ""))
+         addAssert p md (Crucible.SimError (cs ^. MS.csLoc) (Crucible.AssertFailureSimError ("null-equality " ++ MS.stateCond prepost) ""))
 
     MS.SetupGlobal empty _ -> absurd empty
 
@@ -659,7 +641,7 @@ matchTerm sc cc md prepost real expect =
        _ ->
          do t <- liftIO $ scEq sc real expect
             p <- liftIO $ resolveBoolTerm (cc ^. jccSym) t
-            addAssert p md (Crucible.SimError loc (Crucible.AssertFailureSimError ("literal equality " ++ stateCond prepost) ""))
+            addAssert p md (Crucible.SimError loc (Crucible.AssertFailureSimError ("literal equality " ++ MS.stateCond prepost) ""))
 
 ------------------------------------------------------------------------
 
@@ -756,10 +738,6 @@ learnPointsTo opts sc cc spec prepost pt =
 
 ------------------------------------------------------------------------
 
-stateCond :: PrePost -> String
-stateCond PreState = "precondition"
-stateCond PostState = "postcondition"
-
 -- | Process a "crucible_equal" statement from the precondition
 -- section of the CrucibleSetup block.
 learnEqual ::
@@ -776,7 +754,7 @@ learnEqual opts sc cc spec md prepost v1 v2 =
   do val1 <- resolveSetupValueJVM opts cc sc spec v1
      val2 <- resolveSetupValueJVM opts cc sc spec v2
      p <- liftIO (equalValsPred cc val1 val2)
-     let name = "equality " ++ stateCond prepost
+     let name = "equality " ++ MS.stateCond prepost
      let loc = MS.conditionLoc md
      addAssert p md (Crucible.SimError loc (Crucible.AssertFailureSimError name ""))
 
@@ -794,7 +772,7 @@ learnPred sc cc md prepost t =
      u <- liftIO $ scInstantiateExt sc s t
      p <- liftIO $ resolveBoolTerm (cc ^. jccSym) u
      let loc = MS.conditionLoc md
-     addAssert p md (Crucible.SimError loc (Crucible.AssertFailureSimError (stateCond prepost) ""))
+     addAssert p md (Crucible.SimError loc (Crucible.AssertFailureSimError (MS.stateCond prepost) ""))
 
 ------------------------------------------------------------------------
 
@@ -986,11 +964,11 @@ instantiateSetupValue sc s v =
     MS.SetupTerm tt                   -> MS.SetupTerm <$> doTerm tt
     MS.SetupNull ()                   -> return v
     MS.SetupGlobal empty _            -> absurd empty
-    MS.SetupStruct empty _ _          -> absurd empty
+    MS.SetupStruct empty _            -> absurd empty
     MS.SetupArray empty _             -> absurd empty
     MS.SetupElem empty _ _            -> absurd empty
     MS.SetupField empty _ _           -> absurd empty
-    MS.SetupCast empty _ _            -> absurd empty
+    MS.SetupCast empty _              -> absurd empty
     MS.SetupUnion empty _ _           -> absurd empty
     MS.SetupGlobalInitializer empty _ -> absurd empty
   where