+++ saw-remote-api/python/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Revision history for saw-client
 
+## next -- TBA
+
+* Add Python bindings for SAW's experimental MIR verification support:
+
+  * The `mir_load_module` function loads a MIR JSON file into SAW.
+  * The `mir_verify` function performs verification of a MIR function.
+  * The `saw_client.mir` module contains utility functions for constructing
+    MIR types.
+
+  For more information about how SAW's MIR verification support works in
+  general, see the `mir_*` commands documented in the [SAW
+  manual](https://github.com/GaloisInc/saw-script/blob/master/doc/manual/manual.md).
+* The `array()` function now takes an additional `element_type` argument, which
+  defaults to `None`. If constructing a MIR array with no elements, then the
+  `element_type` must be specified. Otherwise, this argument is optional.
+
 ## 1.0.1 -- YYYY-MM-DD
 
 * Add `solver_cache.py` implementing an interface for interacting with SAW